vo_opengl: fix/tweak smoothmotion
This reverts some of the stuff and makes smoothmotion behave a bit more
like current master:

1. The queue is smaller
2. The queue offset is significantly smaller
3. We render new frames before blending

This is justified because the "render after blending" thing doesn't seme
to work consistently any way (introduces stutter due to the way vsync
timing works, or something), so this behavior is a bit closer to master
and makes pausing/unpausing less weird/jumpy.

Framestepping still seems to be affected in some nontrivial way, though.